Dev log space computation

+++ include/kernels/outside.cuh
@@ -7,6 +7,7 @@
 
 #include <cstdint>
 #include <vector>
+
 #include "utils/data_accessing.hpp"
 #include "grammar/grammar.hpp"
 #include "macros.def"
@@ -15,9 +16,9 @@
 #ifdef USE_CUDA
 __global__
 #endif
-void kernel_outside_main(long double* mu, long double* beta, 
+void kernel_outside_main(double* mu, double* beta, 
                         const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
