+++ src/algorithms/alg_inside_outside_main.cpp
@@ -1,10 +1,10 @@
 #include "algorithms/alg_inside_outside_main.h"
 
-long double* outside_algorithm(long double* mu, long double* beta, const uint32_t* sequence, 
+double* outside_algorithm(double* mu, double* beta, const uint32_t* sequence, 
                         uint32_t* pretermination_lookuptable, 
                         uint32_t* grammar_index, 
                         uint32_t* grammar_table, 
-                        long double* alpha, 
+                        double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
@@ -19,11 +19,11 @@ long double* outside_algorithm(long double* mu, long double* beta, const uint32_
     return beta;
 }
 
-long double* em_algorithm_calculate_expection_count(long double* count, long double* mu, long double* beta, const uint32_t* sequence, 
+double* em_algorithm_calculate_expection_count(double* count, double* mu, double* beta, const uint32_t* sequence, 
                         uint32_t* pretermination_lookuptable, 
                         uint32_t* grammar_index, 
                         uint32_t* grammar_table, 
-                        long double* alpha, 
+                        double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
                         #ifdef DEBUG_INSIDE_ALGORITHM
                         ,pcfg* grammar
@@ -41,11 +41,11 @@ long double* em_algorithm_calculate_expection_count(long double* count, long dou
     return count;
 }
 
-long double* inside_algorithm(const uint32_t* sequence, 
+double* inside_algorithm(const uint32_t* sequence, 
                         uint32_t* pretermination_lookuptable, 
                         uint32_t* grammar_index, 
                         uint32_t* grammar_table, 
-                        long double* alpha, 
+                        double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path, pcfg* grammar = nullptr){
